[RENAME] return_foreign to return_foreigns
